+++ lib/ui/elements/matrix.dart
@@ -1,15 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_survey_js/ui/custom_scroll_behavior.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive_nested_form.dart';
+import 'package:flutter_survey_js/ui/survey_configuration.dart';
 import 'package:flutter_survey_js_model/flutter_survey_js_model.dart' as s;
 import 'package:reactive_forms/reactive_forms.dart';
 
 import 'question_title.dart';
 
-Widget matrixBuilder(context, element, {bool hasTitle = true}) {
+Widget matrixBuilder(context, element, {ElementConfiguration? configuration}) {
   return MatrixElement(
     formControlName: element.name!,
     matrix: element as s.Matrix,
-  ).wrapQuestionTitle(element, hasTitle: hasTitle);
+  ).wrapQuestionTitle(context, element, configuration: configuration);
 }
 
 class MatrixElement extends StatelessWidget {
@@ -63,14 +65,19 @@ class MatrixElement extends StatelessWidget {
               ]));
         });
 
-        return Table(
-          defaultColumnWidth: const IntrinsicColumnWidth(),
-          border: TableBorder.all(
-            width: 1.0,
-          ),
-          // columnWidths: map,
-          children: list,
-        );
+        return ScrollConfiguration(
+            behavior: CustomScrollBehavior(),
+            child: Scrollbar(
+                child: SingleChildScrollView(
+                    scrollDirection: Axis.horizontal,
+                    child: Table(
+                      defaultColumnWidth: const IntrinsicColumnWidth(),
+                      border: TableBorder.all(
+                        width: 1.0,
+                      ),
+                      // columnWidths: map,
+                      children: list,
+                    ))));
       }),
     );
   }