+++ lib/ui/elements/multiple_text.dart
@@ -2,12 +2,15 @@ import 'package:built_collection/built_collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_survey_js/ui/form_control.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive_nested_form.dart';
+import 'package:flutter_survey_js/ui/survey_configuration.dart';
 import 'package:flutter_survey_js_model/flutter_survey_js_model.dart' as s;
+import 'package:reactive_forms/reactive_forms.dart';
 
 import 'question_title.dart';
-import 'survey_element_factory.dart';
+import '../survey_element_factory.dart';
 
-Widget multipleTextBuilder(context, element, {bool hasTitle = true}) {
+Widget multipleTextBuilder(context, element,
+    {ElementConfiguration? configuration}) {
   final e = element as s.Multipletext;
   final texts = (e.items?.toList() ?? []).map(toText).toList();
   return ReactiveNestedForm(
@@ -17,7 +20,9 @@ Widget multipleTextBuilder(context, element, {bool hasTitle = true}) {
         shrinkWrap: true,
         itemCount: texts.length,
         itemBuilder: (BuildContext context, int index) {
-          final res = SurveyElementFactory().resolve(context, texts[index]);
+          final res = SurveyConfiguration.of(context)!
+              .factory
+              .resolve(context, texts[index]);
           return index == 0
               ? Padding(
                   padding: EdgeInsets.only(top: 8.0),
@@ -26,16 +31,18 @@ Widget multipleTextBuilder(context, element, {bool hasTitle = true}) {
               : res;
         },
         separatorBuilder: (BuildContext context, int index) {
-          return SurveyElementFactory().separatorBuilder(context);
+          return SurveyConfiguration.of(context)!.separatorBuilder(context);
         },
-      ).wrapQuestionTitle(element, hasTitle: hasTitle));
+      ).wrapQuestionTitle(context, element, configuration: configuration));
 }
 
-Object? multipleTextControlBuilder(s.Elementbase element,
-    {validators = const []}) {
+AbstractControl multipleTextControlBuilder(
+    BuildContext context, s.Elementbase element,
+    {validators = const [], Object? value}) {
   final e = element as s.Multipletext;
   final texts = (e.items?.toList() ?? []).map(toText).toList();
-  final res = elementsToFormGroup(texts, validators: validators);
+  final res = elementsToFormGroup(context, texts,
+      validators: validators, value: e.defaultValue?.value ?? value);
   return res;
 }