+++ lib/ui/elements/panel_dynamic.dart
@@ -2,17 +2,19 @@ import 'package:flutter/material.dart';
 import 'package:flutter_survey_js/ui/form_control.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive_nested_form.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive_nested_group_array.dart';
+import 'package:flutter_survey_js/ui/survey_configuration.dart';
 import 'package:flutter_survey_js_model/flutter_survey_js_model.dart' as s;
 import 'package:reactive_forms/reactive_forms.dart';
 
 import 'question_title.dart';
-import 'survey_element_factory.dart';
+import '../survey_element_factory.dart';
 
-Widget panelDynamicBuilder(context, element, {bool hasTitle = true}) {
+Widget panelDynamicBuilder(context, element,
+    {ElementConfiguration? configuration}) {
   return PanelDynamicElement(
     formControlName: element.name!,
     element: element as s.Paneldynamic,
-  ).wrapQuestionTitle(element, hasTitle: hasTitle);
+  ).wrapQuestionTitle(context, element, configuration: configuration);
 }
 
 class PanelDynamicElement extends StatelessWidget {
@@ -28,6 +30,7 @@ class PanelDynamicElement extends StatelessWidget {
     createNew() {
       //create new formGroup
       return elementsToFormGroup(
+          context,
           (element.templateElements?.map((p0) => p0.realElement).toList() ?? [])
               .toList());
     }
@@ -46,12 +49,15 @@ class PanelDynamicElement extends StatelessWidget {
                     shrinkWrap: true,
                     itemCount: element.templateElements?.length ?? 0,
                     itemBuilder: (BuildContext context, int index) {
-                      final res = SurveyElementFactory().resolve(context,
-                          element.templateElements![index].realElement);
+                      final res = SurveyConfiguration.of(context)!
+                          .factory
+                          .resolve(context,
+                              element.templateElements![index].realElement);
                       return res;
                     },
                     separatorBuilder: (BuildContext context, int index) {
-                      return SurveyElementFactory().separatorBuilder(context);
+                      return SurveyConfiguration.of(context)!
+                          .separatorBuilder(context);
                     },
                   )));
         });