+++ lib/ui/elements/text.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive.dart';
 import 'package:flutter_survey_js/ui/reactive/reactive_color_picker.dart';
+import 'package:flutter_survey_js/ui/survey_configuration.dart';
 import 'package:flutter_survey_js_model/flutter_survey_js_model.dart' as s;
 import 'package:reactive_date_time_picker/reactive_date_time_picker.dart';
 import 'package:reactive_forms/reactive_forms.dart';
 
-import 'question_title.dart';
-
-Widget textBuilder(context, element, {bool hasTitle = true}) {
+Widget textBuilder(context, element, {ElementConfiguration? configuration}) {
   final e = element as s.Text;
   final String? hintText = e.placeholder;
   Widget widget = ReactiveTextField(
@@ -73,25 +72,35 @@ Widget textBuilder(context, element, {bool hasTitle = true}) {
       decoration: InputDecoration(hintText: hintText),
     );
   }
-  return widget.wrapQuestionTitle(element, hasTitle: hasTitle);
+  return widget.wrapQuestionTitle(context, element,
+      configuration: configuration);
 }
 
-Object? textControlBuilder(s.Elementbase element,
-    {validators = const <ValidatorFunction>[]}) {
+AbstractControl textControlBuilder(BuildContext context, s.Elementbase element,
+    {validators = const <ValidatorFunction>[], Object? value}) {
   final e = element as s.Text;
   if (e.inputType == s.TextInputType.date ||
       e.inputType == s.TextInputType.datetimeLocal) {
-    return FormControl<DateTime>(validators: validators);
+    return FormControl<DateTime>(
+        validators: validators,
+        value: e.defaultValue.tryCastToDateTime() ?? value.tryCastToDateTime());
   }
   if (e.inputType == s.TextInputType.color) {
-    return FormControl<String>(validators: validators);
+    return FormControl<String>(
+        validators: validators,
+        value: e.defaultValue.tryCastToString() ?? value.tryCastToString());
   }
   if (e.inputType == s.TextInputType.email) {
-    return FormControl<String>(validators: [...validators, Validators.email]);
+    return FormControl<String>(
+        validators: [...validators, Validators.email],
+        value: e.defaultValue.tryCastToString() ?? value.tryCastToString());
   }
   if (e.inputType == s.TextInputType.number) {
     return FormControl<num>(
-        validators: [...validators, NullableNumberValidator().validate]);
+        validators: [...validators, NullableNumberValidator().validate],
+        value: e.defaultValue.tryCastToNum() ?? value.tryCastToNum());
   }
-  return FormControl<String>(validators: validators);
+  return FormControl<String>(
+      validators: validators,
+      value: e.defaultValue.tryCastToString() ?? value.tryCastToString());
 }