+++ schema/flutter_survey_js_model/CHANGELOG.md
@@ -1,1 +1,19 @@
+## 0.0.2-dev.4
+
+ - revert sdk upgrade
+
+## 0.0.2-dev.3
+
+ - **FIX**: model correctAnswer. ([763dda2b](https://github.com/Goxiaoy/flutter_survey_js/commit/763dda2b8f91f13e87f7f624f94c98b50d826929))
+ - **FIX**: jsonobject value parse. ([fcd048af](https://github.com/Goxiaoy/flutter_survey_js/commit/fcd048af1025bcb66f5dc59f85743bf1b640d759))
+ - **FIX**: multiple_text default value missing. ([d386a2af](https://github.com/Goxiaoy/flutter_survey_js/commit/d386a2af1c8e8736a9ef05e89df3aaab399b18fe))
+
+## 0.0.2-dev.2
+
+ - **FIX**: treat showQuestionNumbersclose null as true.  close [#65](https://github.com/Goxiaoy/flutter_survey_js/issues/65). ([a35afd5f](https://github.com/Goxiaoy/flutter_survey_js/commit/a35afd5f25f245eb97fd9beb78ba09a86b3779be))
+ - **FEAT**: add default value. ([b222b725](https://github.com/Goxiaoy/flutter_survey_js/commit/b222b7259fdc94ed817d8ad205f1d1afa53de437))
+ - **FEAT**: customize configuration use inherited widget. close [#27](https://github.com/Goxiaoy/flutter_survey_js/issues/27). ([e2cb5f11](https://github.com/Goxiaoy/flutter_survey_js/commit/e2cb5f11f6fbe63fbf3eba8e454e7f3941fb4493))
+ - **FEAT**: upgrade deps. update actions. ([1c77db97](https://github.com/Goxiaoy/flutter_survey_js/commit/1c77db97f33a30cb16378803ce2b11e9283063ec))
+ - **FEAT**: upgrade schema. ([af8ac90b](https://github.com/Goxiaoy/flutter_survey_js/commit/af8ac90bafb4f24bca7c4ad17b8f2013531e8824))
+
 ## [0.0.2-dev.1] - 2023/04/26